Remove the wording "sub-350 modes" used in conjunction with VGA double-scanning
The situation is a bit more involved; e.g. the 256-colour and 16/15-bit
512x384 VESA modes on S3 cards are double-scanned, while the 16-colour
and 24/32-bit variants of the same resolution are *not*.

Programs can also set up custom modes that have less than 350 lines, but
are single-scanned, e.g. the 376x308 @ 86Hz mode of the RealSpec ZX
Spectrum emulator, or the 384x282 @ 95Hz mode of the CCS64 emulator.
Both are single-scanned to emulate scanlines.

This is in preparation for the more accurate double-scan detection
approach that relies on the actual contents of the VGA registers that
control double-scanning.

These very special edge cases might seem unimportant, but the more
accurate approach is actually not more difficult to implement, and it
will likely result in more reliable emulation of weird VGA modes across
the board. Additionaly, this will enable us to remove some existing
hacks/workarounds.