Fix height of 200-line CGA, Tandy and PCjr text modes
These erroneously used a value of 400 for all 200-line text modes.
That's clearly wrong as these graphics standards don't have modes with
more than 200 lines of vertical resolution. Things just worked out fine
by fluke as the calculations in vga_draw.cpp are based on the display
timing values which appear to be correct.

But we need the height value to be correct for the new "custom video
mode" detection algorithm that checks whether the actual (calculated)
and "nominal" video mode dimensions differ.