Make mixer channel settings truly persistent
Mixer channel settings are now persisted in a map by their name whenever
a channel gets destroyed or added. This is necessary to make channel
settings "survive" removing an audio device and adding it back later, or
special scenarios such as changing the `sbtype` setting which always
deletes and immediately recreates the SB channel plus optionally the CMS
and OPL channels.