Applied the patch from dosbox-x that daum used.
These are the files copied from doxbox-x commit:
5ec6c54341d148d5e2fda49944f9ed6793538ff8

There's a ton of delted files, but the history of those are in the
dosbox-x repo. trying to get a high signal to noise ratio here.

Eben Bruyns <eben@sdk.co.nz> did the restoration work on this
any eroneous attribution is entirely his fault.

This commit is an attempt to create a clear dilineation between SVN
Dosbox-X and Daum. With the exception of new files This would have
been the starting point for ykhwong in this repo:
https://github.com/ykhwong/dosbox-svn-daum

All code is attributed to Jonathan Campbell <jonathan@castus.tv>