These changes appears to do nothing
I needed these changes to matches the files that I needed for bit
parrity builds. don't ask me why but some of these files are imporant
to have intact for that.

If I had to guess with a gun to my head I would suggest that the
differenes is due to file encodings. I spent hours chasing down a 2 byte
variance in the build cased by src/gui/sdl_mapper.cpp once I included
the file with app the white space changes it worked, the remaining
change that I originally checked in appeared to be a single newline
added to the file between functions.

Eben Bruyns <eben@sdk.co.nz> did the restoration

All code is attributed to:
ykhwong <ykhwong@daum.net>