non functional, but significant changes
These are mostly comments, commented out code and exatra lines or
deleted lines. They should not affect the buidl at all.

I would consider this ykhwongs "finger prints" it highlights his codings
style and commenting habbits. It might be useful to someone trying to ID
his code in other parts of the code bases where it was already absorbed.

If you take the files in this commit and overlay it with the commit I
consider to be his contributions you'll have an almost 100% chance of
having just his work in that set.

Eben Bruyns <eben@sdk.co.nz> did the restoration

all code attributed to:
ykhwong <ykhwong@daum.net>