Changes made to get the code building
These are the changes I made to get the build working for flatpaks.
I tried to keep the changes to a mimimum to avoid instroducing undesired
side effects. I commented out the load and save state code because it
wouldn't work on linux anyway, but this will have to be restored for a
windows build. Another change to be aware of is the src/hardware/ide.cpp
I changed ifdef to ifndef just to make it work on linux, but this will
have to be fixed up properly if a windows build is required.